test(icon): improve cdn prefix testing
